remove obsolete iswitchb support
In version 24.4 of Emacs iswitchb will be obsolete and so we
should remove support at some point; let's get it over with.

Besides if I keep post phoning the Magit release, then Emacs
will beat me to it anyway :-(

If someone actually complains, then we can restore this in
`magit-extra.el'.