remove with-git-editor macro
Remove `with-git-editor'.  It's just like `with-editor' but defaults
to a different environment variable.  Using the latter and explicitly
specifying GIT_EDITOR is not really any worse.  The former also did
bind `magit-process-popup' to -1 but since that is specific to Magit,
which furthermore only used this macro once, this is overkill and
would just complicate the documentation.