remove option magit-use-overlays
Make `magit-insert' and `magit-put-face-property' always use overlays.
Previously it was possible to use text properties which theoretically
gives a small performance boast.  But the option to do the latter was
more confusing than helpful and I doubt that anybody actually made use
of it, especially because doing so would force the use of something
other than the background color for highlighting the current section.