remove option magit-section-highlight-face
This option served to purposes:

It allowed disabling highlighting of the current section.  The ability
to do this was added because of performance considerations, not as an
aesthetic choice.  The effect on performance is not noticeable and so
it is okay to drop support for this.  It is still possible to have the
visual effect of not highlighting the current section by unsetting all
face properties or setting these property to the same value it has in
the `default' face.

Once a theme has set a face property it is not possible for user to
unset that property again, she can just change it to something else,
but not nothing.  This was a problem until recently, but now this can
be worked around by setting that property to the same value it has in
the `default' face.