rearrange section utilities
Create just one section "Section Utilities" by absorbing the directly
preceding, one-definition section "Section Utilities" and moving some of
the definitions previously in "Section Actions" here.  The latter is the
reason we are doing this; we have to move these things up so that we can
use the `magit-section-case' macro in code which is defined earlier than
where it used to be defined.  The `magit-{*}-at-point' functions are
moved elsewhere, putting them below "Section Api" was a mistake in the
first place.