rename faces magit{-diff => }-{file,hunk}-head{er => ing}
Rename `magit-diff-file-header' to `magit-file-heading', and
rename `magit-diff-hunk-header' to `magit-hunk-heading'.

Drop the "diff" from the name of "file" face because it is also used
outside of diffs, i.e. for untracked files.  Drop it from the name of
the "hunk" face too for consistency.

Replace "header" with "heading" because we only use the term "header"
for those one-line headers that appear at the top of some buffers.  For
the part of a section that remains visible even in collapsed state we
use "heading".  Consistently using distinct terms helps avoid confusion.