use magit-diff-file-heading only for text, not complete line
For most headings the heading face is only used for the actual text but
not for the following newline.  This makes a difference when the
background color is set for the heading face.

For hunk headings we also propertize the newline and by default also set
the background color, which makes diffs more readable because this
clearly separates different hunks from one another.

Until now we also did that for files because a diff may contain several
files and separating them clearly makes sense the same way it does for
hunks.  We did not actually set the background color by default, but
users who wanted to visually separate files the same way that hunks are,
just had to do so.

However if the background color actually is specified and spans the hole
line then the region is not visible when selecting multiple files.  This
would be very confusing so remove the ability to set the background for
the complete file heading line.