use shade of grey for magit-hash
Hashes should not stand out because, from a human perspective, they are
usually not very interesting.  So use some shade of grey instead, to
make them *less* visible.  Previously some red was used which was also
semantically wrong; red is the color used for e.g. *removed* lines,
while the main even in a commit's life is its *creation*.