use shades of grey for highlighting and refname faces
Previously `magit-section-highlight' inherited from
`secondary-selection'.

Doing so had the advantage that for most themes this face is
sufficiently distinct form other faces without requiring any further
customization.  The default background color for `secondary-selection'
was selected so that it works regardless of what changes were made to
other faces.  This is done by using a background color which unlikely
to be used as background or foreground color for any face in most
reasonable themes.  The disadvantage, and result, of this is that
this face always looks very out of place.

For `secondary-selection' this works (not the "secondary").  But for
`magit-section-highlight' it doesn't; we use that face all the time,
and we want Magit to look okay when using the default faces (dark or
light).