define magit-diff-{added,removed,context} faces from scratch
No longer inherit from `diff-{added,removed,context}' because these
faces only define the background color.  It is better to also define the
foreground color because other faces that overlay the text propertized
with these faces might also set the background color (which would shadow
the background colors of the diff faces, but not their foreground colors).

This is specially problematic because in Magit the current section is,
by default, highlighted using the background color.  That will be fixed
in the next commit, but the same applies to the region.