generalize "oneline" log style
Generalize the "oneline" log style, so that it can be used in placed
where previously the "unique" style was used.

Create the new functions `magit-insert-log' and `magit-insert-log-long'
from code previously in `magit-refresh-log-buffer'.

Because we do not want to show ref names in log sections (as opposed to
buffers) by default, remove "%d" from the default `--format' and instead
add `--decorate' as a popup argument and enable it by default.

This argument is a special case; if it is present it is replaced with
`--decorate=full', because we need full refs to be able to reliably
parse them.  But we then shorten them similar to how Git would do it,
with `--decorate[=short]'.  This is an implementation detail so we only
tell users about `--decorate'.