always fallback to magit-checkout-stage when discarding
When the user "discards" a conflict, then always offer to a side using
`magit-checkout-stage'.  Previously we only did so when acting on a file
section that was listed below the "unstaged" changes, but not when it
was listed below "staged".  Because these section really refer to the
same conflict, it does not make sense to redirect the reasonable action
in one case, while throwing up our hands in the other.

To allow this change (and others to follow) move most of the work
involved in discarding an individual file to `magit-discard-file'.
Previously it was randomly split across that function and
`magit-discard'.