prepare the log margin upfront
Previously, when the margin was disabled, we did not create the overlays
that make up the margin to avoid unnecessary work.  That work is not
very involved so we did not save much.  What is involved though is
having to recreate the complete log buffer when later enabling the
margin; and so it is better to do this right away.