(magit-refresh-commit-buffer): add --find-renames
Currently, when a file is renamed, showing that commit in magit results
in two separate files: a fully deleted file where every line is a
removal and a fully added file where every line is an addition. This is
usually just spammy and not useful. Leverage git's --find-renames option
to git-log (and friends) which makes the output much more succinct and
readable.