for pcase, require at least version 24.1 of Emacs
Sooner or later this had to happen.  While I would have preferred to
delay this move until after releasing version 2.1.0, it was always
clear, that I was only willing to maintain backward compatibility as
long as the cost was not very high.

Denying myself the use of `pcase' is such a cost.  I am going to use
this macro a lot, and while it would be possible to implement the
things to come without using `pcase', doing so would not be pleasant.
So unpleasant in fact, that the temptation would be big to write a
specialized "case" macro, similar to `magit-section-case'.  But my
experience with that, and `magit-section-action', has taught me that
going down that road leads to over-specialized abstractions which
eventually end up being rather featureful, but never-the-less to
restrictive even for its intended use-cases.