allow discarding just the un-/staged part of a file
Previously it was only possible to discard a file if it had either
staged or unstaged changes.  Remove that limitation.  Note that
discarding staged changes is done in two steps by first unstaging
and then applying the same changes in reverse to the worktree.
This second step can fail due to conflicts.