improve discarding of changes
I.e. (1) clearly separate it from "reversing" and "unstaging" and
(2) teach it about all the change units and sets a user might want
to discard.

Git doesn't know how to discard a change, so we have to use various
commands to implement this.  Require confirmation from the user for
each discard variant individually.  The variants are `delete' an
added file, `resurrect' a deleted file, undo a `rename', `discard'
an edit, and `resolve' a conflict by picking a side (and therefor
discarding the other).