rename magit-revert to magit-reverse
While both "revert" and "reverse" are used for some kind of "undo",
these actions are only somewhat related and definitely not the same.

While a `revert' Git command exists, there is no `reverse' command.
That term appears in form of the `--reverse' argument to the `apply'
command.

Because it is confusing that the command `magit-revert' does *not*
use or do the same thing as the command `git revert', rename it to
`magit-reverse', which makes sense because it ultimately ends up
using `git apply --reverse'.

Also move the definition of `magit-revert-commit' (which is Magit's
wrapper around `git revert') away from the `magit-reverse'.