magit-reverse: extend scope
Teach `magit-reverse' how to reverse multiple files (region and list),
at once.  Also teach it when reversing doesn't make sense: when the
change isn't tracked by Git throwing it away is enough.  The correct way
to make an unstaged change go away is to discard it, the correct way to
make an untracked file go away is to delete it (which is a variant of
discard).

Add new functions `magit-reverse-files` and `magit-reverse-apply', to
implement this.