add magit-refs-mode and magit-show-refs
Add a new mode `magit-refs-mode' which replaces the two old modes
`magit-branch-manager-mode' and `magit-wazzup-mode'.  The new entry
command is `magit-show-refs', replacing `magit-branch-manager' and
`magit-wazzup'.

Originally these had to be different modes because wazzup was so
slow that users would only tolerate it when they actually needed the
additional information it provided compared to the branch manager.
I have made wazzup fast quite some time ago but until now didn't get
around to merging the two modes.

The new mode is called "Refs Mode" because it lists all kinds of
references not just branches (that was already the case of the old
"'Branch' Manager" too).