change key bindings for magit-show-refs and magit-cherry
The old `magit-branch-manager' command was only bound in popups while
the old `magit-wazzup' command was only bound in the mode map (and the
dispatch popup).

Since there now is only `magit-show-refs', we can do better.  Because
this command is important enough give it the mode map.  But don't use
"w" which was used by `magit-wazzup' (and which we are going to use for
one of the sequence commands) instead use "y".  "y" was previously bound
to `magit-cherry', which can be considered a lightweight, but less
important, variant of the cherry-feature of `magit-show-refs', and is
now moved to "Y".

Do not bind `magit-show-refs' in any popups (except for the dispatch
popup).  Commands that are bound in the mode map don't need to be
available from popups (except the dispatch popup whose purpose is to
help users who don't know the binding in the mode map).