magit-ediff-resolve-file: new command
Add `magit-ediff-resolve-file', a simple wrapper around `smerge-ediff',
as a replacement for `magit-interactive-resolve'.  This fixes #256, i.e.
only unresolved commits are shown.  Ediff was created in the age of vcs
that are bad at merging and has not yet been adjusted for dvcs like Git
and Mercurial, and so it actually wants to do all everything that needs
to be done to do a merge.  However we want to initiate the merge using
Git and then we want Ediff to help us with the conflicts Git could not
automatically resolve; we don't want to also have Ediff attempt to
resolve the conflicts that Git already resolved, especially because
Ediff fails more often than Git, causing the user additional work.
That's why the command name contains "resolve" not "merge", we only
use Ediff to resolve conflicts, not to do a merge.