magit-paint-hunk: don't paint typechanges
A typechange is treated as a hunk because it can be staged like a hunk
and because it saves a lot of work if we don't have to explicitly teach
many commands how to handle typechanges.  But `magit-paint-hunk' only
knows (and needs to) paint real hunks, so make it stop attempting to do
so for typechanges.

Close #1424.