magit-update-index: new command
Note that this command and `magit-find-file-index-noselect' are only
intended to be used with Magit's upcoming Ediff integration.  They are
written in a way that they can also be used elsewhere, but if you do,
then you will surely run into problems; there are many details this
simple implementation does not take care of.