magit-read-file-from-rev: Fix default-directory
The default directory was not set when `magit-revision-files' was
called, resulting in the files being limited to the current directory.
Aside from the unintentional limiting of files, this resulted in
`magit-find-file{,-other-window}' displaying a blank buffer when called
from a subdirectory (because the file path was not correct for `git
cat-file').