magit-blame-mode: bugfix
Users should use `magit-blame' or `magit-blame-popup' to turn on
`magit-blame-mode', but the intention was to also support using
the mode command for this purpose.  For that to work we have to
bind `magit-blame-mode' to nil around the call to `magit-blame',
because that otherwise thinks we are doing an recursive blame.