magit-revert-buffers: don't use auto-revert-handler
`auto-revert-handler' does quite a few things we don't need, so instead
of using that just call `revert-buffer', `vc-find-file-hook' directly.

Since that also means that our advice to `auto-revert-handler' is
no longer used also turn `magit-blame-mode' off and on again when
appropriate.  But do not remove the advice, it is still needed when
`auto-revert-handler' is called outside of our control.