sanitize preparation of arguments to Git
Add a single function `magit-process-git-arguments' which prepares
the arguments to Git.  This includes flattening the list of arguments,
prepending `magit-process-standard-options', and when Cygwin Git is used
the quoting of certain problematic strings.

All of this is now only done for Git processes, previously we did also
flatten arguments in `magit-call-process' and `magit-start-process'.
Also avoid repeatedly flattening the arguments, by only doing so at the
lowest possible level, in calls to built-in process functions, i.e. in
`magit-git-exit-code', `magit-git-string', `magit-git-insert',
`magit-git-lines', `magit-run-git, and `magit-start-git'.

Arguments are now fully flattened.  Previously only the first level was
flattened.  Or at least that is what the doc-strings claimed, in reality
a varying number of levels was flattened (because the "one-level only"
flattening happened several times).

Because the argument list is now fully flattened - just before calling
Git - we can also drop several uses of `apply'.

Remove the option `magit-process-quote-curly-braces' (for Cygwin Git)
and instead put the test previously used to set its value directly into
`magit-process-git-arguments'.  While there never is absolute certainty
with things concerning Windows, there also have been no reports about
the quoting happening when it shouldn't, or not happening when it
should.  So remove the option to avoid confusing users (also the
doc-string was not really correct).