magit-blame-process-filter: avoid empty lines
When splitting partial output from `git blame --incremental' drop empty
lines, i.e. very partial output which we cannot process in a meaningful
way.  There are not actually empty lines in the output, but we might
receive partial output which only consists of a newline.  Dropping the
newline does not cause any problems latter when we combine partial
output, because we store partial output as a list of lines.