with-editor-process-filter: expand relative file name
The child process might call the $EDITOR with a relative file name,
so we have to remember the working directory when the process is
created and then, if we get a relative file name, expand it in
`with-editor-process-filter'.

The easiest way to pass the `default-directory' from our advice to
`start-file-process' to `with-editor-process-filter', is to store it
in the process object.