fontify comments in the regexp pass
In Git commit messages comments are lines beginning with # (actually
`core.commentchar').  A # later in a line does not begin a comment,
so it is common to use #N to refer to the issue N.

Until now # and everything that follows was always fontified as
comment because we used syntactic fontification.

We already defined a keyword regexp for comments, but that was
redundant.  We also set `comment-use-syntax' to nil, but that only
affects the commands in `newcomment', not actual fontification.

We now turn off syntactic fontification by not adding `core.commentchar'
to the syntax table as a comment starter .  Instead we add it as
punctuation.  (The major mode might already have added it as a comment
starter, we need to override that).

For some reason when a syntactic-pass string appears inside a
regexp-pass comment, then nothing on that line actually is displayed as
comment.  Until we know the cause of this, we work around it by also
defining potential string delimiters as punctuation.