magit-branch-popup: removed unused arguments
Remove the `--merged', `, `--no-merged' and `--contains' arguments from
the `magit-branch-popup'.  This popup does not feature a "list branches"
command and these options are only meaningful with such a command.  Not
only do this options no longer serve any purpose, if a user actually
sets them and then uses one of the commands which actually respect
`magit-current-popup-args' then that results in an error.

Once upon a time these switches were used by `magit-branch-manager'
which no longer is available from this popup, and which not even exists
anymore.  Its successor `magit-show-refs' no longer supports these
arguments. It could possibly make sense to restore support for these
arguments.  If we implement that, then it should be done using a popup
`magit-show-refs-popup', with only one action `magit-show-refs'.