magit-refresh-buffer: do nothing if magit-refresh-function is nil
This is done to avoids unnecessary work.  But it also prevents
`magit-refresh-buffer-hook' from running, which might be a problem for
third-party code which for some reason expects this to run even if the
buffer was not actually refreshed.  If this turns out to be the case
then we could change back to always run that hook.