remove variable magit-status-line-align-to
This variable was used when padding the leading keywords of header lines
in the status buffer.  Its purpose was to allow third-party extensions
that use headers with keywords that are longer than those used in Magit,
to adjust the padding accordingly.  By removing this variable, we remove
that - so far unused - feature.  If needed this can add such a variable
again.

For why we "needlessly" do that, see the next few commits.