magit-commit-add-log: Use symbol boundary in function name REGEXP
Use symbol boundary instead of word boundary in function name REGEXP. This is
because using word boundary causes partial function name being matched. As an
example, let's say we have two diff hunks, one in the function `foo' and the
other in `foo-bar'. When the word boundary REGEXP is used, the REGEXP matches
both function name `foo' and `foo-bar', which will prevent
`magit-commit-add-log' from adding a log entry for function `foo-bar' if an
entry for function `foo' already exist.