remove magit-set-variable-and-refresh
Remove `magit-set-variable-and-refresh' which was used to set
various Custom options and the refresh some Magit buffers using
`magit-refresh-all'.  That no longer worked as expected because
the latter function now only refreshes Magit buffers in the current
repository.  So this function only had an effect in the repository
at ~/.emacs.d.