magit-diff-options: correct custom group
Put `magit-diff-options' in Custom group `magit-popups' instead of
`magit'.  Do *not* put it in `magit-diff' because all equivalent
variables are currently placed in `magit-popups'.

(All of these variables, except for this one, are named
`magit-...-popup-defaults'.  We don't change the name of this variable
because the convention also isn't good.  All of these variables,
including this one, will be renamed at the same time).