add custom group magit-commands
Add a new Custom group for options that control the behavior of certain
commands, and for which no other suitable Custom group exists.

Many of these options tune the behavior of commands that *create*
commits.  These options were previously placed in the `magit-commit'
group, but that group is for functionality related to the inspection
of *existing* commits.  The other options were previously placed in
`magit-modes' group, which is wrong because none of them have directly
to do with any of the various major modes provided by Magit.

An alternative to creating `magit-commands' would have been to place
these options in `magit', but I would prefer to reserve that for options
with "global" effects, the exact opposite of what these the options in
question do.