begin rearranging magit.el
Actually we already began with that some commits ago, but this is the
beginning of the final step.

The goal is to have five top-level sections: Options, Inspect,
Manipulate, Transfer and Miscellaneous.  The main split, which makes it
necessary to move many second level sections is between Inspect,
Manipulate and Transfer.  Because not everything fits perfectly in these
three categories, we also need Miscellaneous.  Even so some things that
will be placed in one of the tree main sections, will not fit perfectly.
Some inconsistency cannot be avoided.

This commit renames Modes to Inspect and moves Porcelain>Visit to
Inspect>Files.  Because Visit was the first subsection of Porcelain and
now is the last subsection of Inspect, we can do so by moving the
Porcelain heading instead of the Visit code.

Also change the names of the subsections of Porcelain, losing the "-ing"
suffix.  Now Porcelain is a misnomer, because other sibling sections are
just as much about porcelain - and just as much *not* about that.  We
cannot split Porcelain into Manipulate and Transfer yet, because the
subsections are not in the right order yet.