turn some sections into subsections of "Miscellaneous"
Move the beginning of the "Miscellaneous" top-level section, so that
the sections "Tag", "Stash" and "Submodules" are now subsections of
"Miscellaneous" instead of "Porcelain".

This would also result in a subsection of the top-level section named
"Miscellaneous" which in turn is also named "Miscellaneous".  Rename
that subsection to "Various" and move it to the end of "Miscellaneous".