magit-insert-branch: set magit-section property for post-children text
In the lambda which is created by `magit-insert-branch' and later used
when inserting and washing the section body when it gets expanded, set
the `magit-section' property of the post-children separator newline.

Setting this property is normally done inside `magit-insert-section',
here we have to do it explicitly because we modify an existing section.